Use PRI?64 instead of "ll?" in format strings (continued).
Continuation of work started in commit 15a79c73, after initial trial.

Author: Thomas Munro <thomas.munro@gmail.com>
Discussion: https://postgr.es/m/b936d2fb-590d-49c3-a615-92c3a88c6c19%40eisentraut.org
(cherry picked from commit a0ed19e0a9efe93b3b83d6e3fe8f77656be253a2)

+++ src/backend/utils/activity/pgstat.c
@@ -1688,9 +1688,9 @@ pgstat_write_statsfile(void)
 		 */
 		if (!pgstat_is_kind_valid(ps->key.kind))
 		{
-			elog(WARNING, "found unknown stats entry %u/%u/%llu",
+			elog(WARNING, "found unknown stats entry %u/%u/%" PRIu64,
 				 ps->key.kind, ps->key.dboid,
-				 (unsigned long long) ps->key.objid);
+				 ps->key.objid);
 			continue;
 		}
 
@@ -1912,9 +1912,9 @@ pgstat_read_statsfile(void)
 
 						if (!pgstat_is_kind_valid(key.kind))
 						{
-							elog(WARNING, "invalid stats kind for entry %u/%u/%llu of type %c",
+							elog(WARNING, "invalid stats kind for entry %u/%u/%" PRIu64 " of type %c",
 								 key.kind, key.dboid,
-								 (unsigned long long) key.objid, t);
+								 key.objid, t);
 							goto error;
 						}
 					}
@@ -1985,9 +1985,9 @@ pgstat_read_statsfile(void)
 					if (found)
 					{
 						dshash_release_lock(pgStatLocal.shared_hash, p);
-						elog(WARNING, "found duplicate stats entry %u/%u/%llu of type %c",
+						elog(WARNING, "found duplicate stats entry %u/%u/%" PRIu64 " of type %c",
 							 key.kind, key.dboid,
-							 (unsigned long long) key.objid, t);
+							 key.objid, t);
 						goto error;
 					}
 
@@ -1998,9 +1998,9 @@ pgstat_read_statsfile(void)
 									pgstat_get_entry_data(key.kind, header),
 									pgstat_get_entry_len(key.kind)))
 					{
-						elog(WARNING, "could not read data for entry %u/%u/%llu of type %c",
+						elog(WARNING, "could not read data for entry %u/%u/%" PRIu64 " of type %c",
 							 key.kind, key.dboid,
-							 (unsigned long long) key.objid, t);
+							 key.objid, t);
 						goto error;
 					}
